Refactor Day4 code for fun
This is a non-functional commit.

Yes, it's just AoC code, but after a few days I thought I'd clean up at
least some of it.

This is a slightly better view of the sort of code I'd like to write.
Obviously, if this were production Rust, I'd want better error handling
and testing, but it was fun to improve the solution to today's problem
none-the-less =)