Adding incomplete day 13 part one
Not going to make it on time!

I've struggled with this problem on two fronts: reading comprehension
and parsing (chronologically).

I misread how many integers must be compared before an ordering decision
can be made. While it did take time to narrow down the bug, it was more
bewildering to stare at the example and still not get it.

Now I'm struggling with parsing. This should be easier. I'll take a look
later with fresh eyes.

Oh well. There were a couple other puzzles I struggled with too. It's
always for different reasons, oddly enough.