Add day 15 part one
This was the first puzzle I had to noodle with. There was some hint of
the geometric side of this problem in the description but it wasn't
obvious to me.

I think the solution I have is nice. I'm sure it could be written with
less, but it's very easy to follow. In fact, it was written in a way
that mimics how I described the problem and solution in my head.

On Reddit, I saw mention of brute force solutions. This probably isn't
brute force. It takes longer for `rustc` to compile `-- release` that it
does to find a solution.